surfaces: factor out common tiled interfaces, code
Add interfaces for things directly accessible, blittable, or
compositable by tile.

Move save_as_png to lib.surface
